Merging 0.5
Deleted Geohandler because it is now merged with History. That way we
can update the arrayadapter. The merge is going well, but there is a
whole slew of new problems (as expected). It seems like when we do get
it working that the app actually saves inscritpions to the display when
it receives them.